Made cost prices ulong and ints

+++ Idle Defense/Assets/Editor/TurretCSVtoSO.cs
@@ -76,13 +76,13 @@ namespace Assets.Editor
             turretInfo.Damage = float.Parse(lines[2], CultureInfo.InvariantCulture);
             turretInfo.DamageLevel = 0;
             turretInfo.DamageUpgradeAmount = float.Parse(lines[3], CultureInfo.InvariantCulture);
-            turretInfo.DamageUpgradeBaseCost = float.Parse(lines[4], CultureInfo.InvariantCulture);
+            turretInfo.DamageUpgradeBaseCost = int.Parse(lines[4], CultureInfo.InvariantCulture);
             turretInfo.DamageCostExponentialMultiplier = float.Parse(lines[5], CultureInfo.InvariantCulture);
 
             turretInfo.FireRate = float.Parse(lines[6], CultureInfo.InvariantCulture);
             turretInfo.FireRateLevel = 0;
             turretInfo.FireRateUpgradeAmount = float.Parse(lines[7], CultureInfo.InvariantCulture);
-            turretInfo.FireRateUpgradeBaseCost = float.Parse(lines[8], CultureInfo.InvariantCulture);
+            turretInfo.FireRateUpgradeBaseCost = int.Parse(lines[8], CultureInfo.InvariantCulture);
             turretInfo.FireRateCostExponentialMultiplier = float.Parse(lines[9], CultureInfo.InvariantCulture);
 
             turretInfo.RotationSpeed = float.Parse(lines[18], CultureInfo.InvariantCulture);
@@ -95,13 +95,13 @@ namespace Assets.Editor
             turretInfo.CriticalChance = float.Parse(lines[10], CultureInfo.InvariantCulture);
             turretInfo.CriticalChanceLevel = 0;
             turretInfo.CriticalChanceUpgradeAmount = float.Parse(lines[11], CultureInfo.InvariantCulture);
-            turretInfo.CriticalChanceUpgradeBaseCost = float.Parse(lines[12], CultureInfo.InvariantCulture);
+            turretInfo.CriticalChanceUpgradeBaseCost = int.Parse(lines[12], CultureInfo.InvariantCulture);
             turretInfo.CriticalChanceCostExponentialMultiplier = float.Parse(lines[13], CultureInfo.InvariantCulture);
 
             turretInfo.CriticalDamageMultiplier = float.Parse(lines[14], CultureInfo.InvariantCulture);
             turretInfo.CriticalDamageMultiplierLevel = 0;
             turretInfo.CriticalDamageMultiplierUpgradeAmount = float.Parse(lines[15], CultureInfo.InvariantCulture);
-            turretInfo.CriticalDamageMultiplierUpgradeBaseCost = float.Parse(lines[16], CultureInfo.InvariantCulture);
+            turretInfo.CriticalDamageMultiplierUpgradeBaseCost = int.Parse(lines[16], CultureInfo.InvariantCulture);
             turretInfo.CriticalDamageCostExponentialMultiplier = float.Parse(lines[17], CultureInfo.InvariantCulture);
         }
 
@@ -110,13 +110,13 @@ namespace Assets.Editor
             turretInfo.PelletCount = int.Parse(lines[10]);
             turretInfo.PelletCountLevel = 0;
             turretInfo.PelletCountUpgradeAmount = int.Parse(lines[11]);
-            turretInfo.PelletCountUpgradeBaseCost = float.Parse(lines[12], CultureInfo.InvariantCulture);
+            turretInfo.PelletCountUpgradeBaseCost = int.Parse(lines[12], CultureInfo.InvariantCulture);
             turretInfo.PelletCountCostExponentialMultiplier = float.Parse(lines[13], CultureInfo.InvariantCulture);
 
             turretInfo.DamageFalloffOverDistance = float.Parse(lines[14], CultureInfo.InvariantCulture);
             turretInfo.DamageFalloffOverDistanceLevel = 0;
             turretInfo.DamageFalloffOverDistanceUpgradeAmount = float.Parse(lines[15], CultureInfo.InvariantCulture);
-            turretInfo.DamageFalloffOverDistanceUpgradeBaseCost = float.Parse(lines[16], CultureInfo.InvariantCulture);
+            turretInfo.DamageFalloffOverDistanceUpgradeBaseCost = int.Parse(lines[16], CultureInfo.InvariantCulture);
             turretInfo.DamageFalloffOverDistanceCostExponentialMultiplier = float.Parse(lines[17], CultureInfo.InvariantCulture);
         }
 
@@ -125,13 +125,13 @@ namespace Assets.Editor
             turretInfo.PierceChance = float.Parse(lines[10], CultureInfo.InvariantCulture);
             turretInfo.PierceChanceLevel = 0;
             turretInfo.PierceChanceUpgradeAmount = float.Parse(lines[11], CultureInfo.InvariantCulture);
-            turretInfo.PierceChanceUpgradeBaseCost = float.Parse(lines[12], CultureInfo.InvariantCulture);
+            turretInfo.PierceChanceUpgradeBaseCost = int.Parse(lines[12], CultureInfo.InvariantCulture);
             turretInfo.PierceChanceCostExponentialMultiplier = float.Parse(lines[13], CultureInfo.InvariantCulture);
 
             turretInfo.PierceDamageFalloff = float.Parse(lines[14], CultureInfo.InvariantCulture);
             turretInfo.PierceDamageFalloffLevel = 0;
             turretInfo.PierceDamageFalloffUpgradeAmount = float.Parse(lines[15], CultureInfo.InvariantCulture);
-            turretInfo.PierceDamageFalloffUpgradeBaseCost = float.Parse(lines[16], CultureInfo.InvariantCulture);
+            turretInfo.PierceDamageFalloffUpgradeBaseCost = int.Parse(lines[16], CultureInfo.InvariantCulture);
             turretInfo.PierceDamageFalloffCostExponentialMultiplier = float.Parse(lines[17], CultureInfo.InvariantCulture);
         }
 
@@ -140,13 +140,13 @@ namespace Assets.Editor
             turretInfo.ExplosionRadius = float.Parse(lines[10], CultureInfo.InvariantCulture);
             turretInfo.ExplosionRadiusLevel = 0;
             turretInfo.ExplosionRadiusUpgradeAmount = float.Parse(lines[11], CultureInfo.InvariantCulture);
-            turretInfo.ExplosionRadiusUpgradeBaseCost = float.Parse(lines[12], CultureInfo.InvariantCulture);
+            turretInfo.ExplosionRadiusUpgradeBaseCost = int.Parse(lines[12], CultureInfo.InvariantCulture);
             turretInfo.ExplosionRadiusCostExponentialMultiplier = float.Parse(lines[13], CultureInfo.InvariantCulture);
 
             turretInfo.SplashDamage = float.Parse(lines[14], CultureInfo.InvariantCulture);
             turretInfo.SplashDamageLevel = 0;
             turretInfo.SplashDamageUpgradeAmount = float.Parse(lines[15], CultureInfo.InvariantCulture);
-            turretInfo.SplashDamageUpgradeBaseCost = float.Parse(lines[16], CultureInfo.InvariantCulture);
+            turretInfo.SplashDamageUpgradeBaseCost = int.Parse(lines[16], CultureInfo.InvariantCulture);
             turretInfo.SplashDamageCostExponentialMultiplier = float.Parse(lines[17], CultureInfo.InvariantCulture);
         }
 
@@ -155,13 +155,13 @@ namespace Assets.Editor
             turretInfo.PercentBonusDamagePerSec = float.Parse(lines[10], CultureInfo.InvariantCulture);
             turretInfo.PercentBonusDamagePerSecLevel = 0;
             turretInfo.PercentBonusDamagePerSecUpgradeAmount = float.Parse(lines[11], CultureInfo.InvariantCulture);
-            turretInfo.PercentBonusDamagePerSecUpgradeBaseCost = float.Parse(lines[12], CultureInfo.InvariantCulture);
+            turretInfo.PercentBonusDamagePerSecUpgradeBaseCost = int.Parse(lines[12], CultureInfo.InvariantCulture);
             turretInfo.PercentBonusDamagePerSecCostExponentialMultiplier = float.Parse(lines[13], CultureInfo.InvariantCulture);
 
             turretInfo.SlowEffect = float.Parse(lines[14], CultureInfo.InvariantCulture);
             turretInfo.SlowEffectLevel = 0;
             turretInfo.SlowEffectUpgradeAmount = float.Parse(lines[15], CultureInfo.InvariantCulture);
-            turretInfo.SlowEffectUpgradeBaseCost = float.Parse(lines[16], CultureInfo.InvariantCulture);
+            turretInfo.SlowEffectUpgradeBaseCost = int.Parse(lines[16], CultureInfo.InvariantCulture);
             turretInfo.SlowEffectCostExponentialMultiplier = float.Parse(lines[17], CultureInfo.InvariantCulture);
         }
     }
